requirements.txt
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
Pillow>=8.0.0
